There are a few more changes in this than I might like in a single commit.
I implemented a TerminalPanel which expects the input to be handled some other method, but has most of the expected terminal control functions. The initial tests showed it not significantly better than the existing approach for increasing performance. It does allow some additional uses, though, so it is still handy. -- Along with this, I have an AwtPalette which accepts the same color mappings but returns the results directly as Awt Colors.

I am officially phasing out using int[] as return types. I created a Sizable/SimpleSize interface/implementation to assist with this.

There's an untested Navier-Stokes implementation. I was going to test it when I got distracted. This features both a simple-line function for bulk operations and an iterator which will walk through the cells one at a time.

I cleaned up the name of some parameters.

ColorHelper has a few tweaks. It can go to/from HSV and has a helper function to saturate/desaturate a whole palette (or a range therein).

This also checks in my first attempt at a Perlin Noise generator. Well, it is mostly comes from Java sample code, so it should work. It hasn't been tested, though. This is the third revision of the code -- the one sometimes called "Simplex Noise". Since it is faster and better looking than the earlier version, this is the only one that will be available here.

Unfortunately, there's probably more than this.