Fix for Issue 3.
Alt-Enter now works properly. It currently needs to be KEY_ENTER and not KEY_NP_ENTER, but I think that's the expected behavior.

It also needs to be _just_ Alt. AltGr doesn't do it. Shift+Alt-Enter doesn't do it. The goal is that applications should rarely need to turn it off.