Massive display speed improvement.
This also adds a set of "V2" classes which exist solely in the trunk as a place to experiment with things which may become the Ver 2.0 code without breaking things.

The display is now synchronous with the main logic and we exclusively rely upon refresh statements to update the window. This is assisted by the standard Swing backing store which redisplays the window with the last-good data when it was obscured.