It mostly seems to work.
This has a lot of dungeon-digging helper crap. Some of it appears to need comprehensive JUnit tests -- at least from some of the issues I've run in to. (If only I'd already resolved most of them -- but that's always a pipe dream if you're not actually writing tests.)

The SimpleDigger.java will probably be refactored to have an AbstractDigger and we'll get a fancier digger that will support cells with a sticky terrain. (As opposed to needing to cache the previous cell contents.)