Issue 18 : serialization of Grid that can withstand version changes robustly
This uses a helper class and serializes that class instead of the Grid class.

The helper class is light-weight and communicates to newer versions of the helper class -- and the main Grid class -- through a simple Map. This allows the Grid to only know about the current version and each version only knows about the one one-version one step newer.

It looks like there may be an opportunity for a common base class and interface, but I was hoping to avoid that and keep strict POJOs. In any case, a base class solution will only work if such a solution will be able to access the fields. (And I wanted to check it in while it works.)