Issue 31: getString function that still allows application to catch and modify events - complete
Do some fairly major refactoring of how the getString() call handles the callbacks. This was initially sketched out when I did the fix for the Blacken-1.0.1 release -- but it was incomplete and rushed. I didn't like the name or location. This is a better long-term solution.

This also brings up the first logical (or command) codepoint instead of of a physical codepoint. Instead of having a KEY_ prefix they have a CMD_ prefix. This will be used further once work on Issue 32 (Logical key commands and standardized alternate keyboard mappings) gets under way.