Issue 29:	By default, line-drawing characters are unusable - resolved
So, DeJa Vu Mono is now checked in. It is mostly public domain, but I'm bundling its license so there's no confusion. It solves the issue with line-drawing characters, but there are some issues with the "SHADE" codepoints. (They seem short.) I don't see how there's a solution until we start tweaking glyphs.

I've heard Eclipse complained that setFont(null) could be confused for either setFont(String) and setFont(String...). I've fixed this by giving setFont(String) a parameter which allows for fonts to be checked for existence without actually setting the font. This doesn't break the ABI because 1.0 didn't have this call. This doesn't break the API because setFont(String...) will be used instead.