Issue 39: We need a reusable pager to display help and license texts - fixed
Adds a simple pager with the StringViewer class. Adds a TermnalView which restricts writes to a subsection of a window.

The Obligations class now has a function to get the name of the bundled font. This is useful when dealing with the font license text.

Various things have been pushed around slightly to make way for the TerminalView class (and the TerminalViewInterface).

BreakableLoop was created to simplify the creation of components which do one thing over time -- maybe even process keystrokes on their own -- but also need to be interruptable. BreakableLoop works with the new Steppable interface to perform get things done.

CodepointCallbackInterface is generally useful, so the one function there that made it look like it was only for editor-like things was renamed.