Fix the fbm PerlinNoise functions to be static.
The existing TerminalInterface didn't have a putString -- it seemed a painful thing to be lacking.

putString -- like getString -- normally doesn't change the attributes. We now have a "TerminalCellTemplate" to assist with overwriting attributes without overwriting the visible codepoint sequence.

TerminalCellTemplate also supports the new TerminalCellTransformer feature. The idea is that functions that produce procedural textures can use this.